feat: Use most compact JSON encoding
This shrinks event sizes a bit, even when gzip'ed.
The compact representation is documented in the json module.

Alternatively, we can also look into using a custom encoder (that could
also handle datetime objects, instead of the current manual
serialization of those).

In the absence of proper benchmark data, consider a random transaction
event t:

>>> len(json.dumps(t)), len(json.dumps(t, separators=(',', ':')))
(82174, 78516)

That is 95.5% of the original size.

With gzip compression:

>>> len(gzips(json.dumps(t))), len(gzips(json.dumps(t, separators=(',', ':'))))
(13093, 12988)

That is 99.2% of the original size.